mod function

+++ GodWork/SQL/FUNCTION/fn_minute_to_string.sql
@@ -1,8 +1,8 @@
--- Function: public.fn_minute_to_string(integer)
+-- Function: public.fn_minute_to_string(bigint)
 
--- DROP FUNCTION public.fn_minute_to_string(integer);
+-- DROP FUNCTION public.fn_minute_to_string(bigint);
 
-CREATE OR REPLACE FUNCTION public.fn_minute_to_string(minute integer)
+CREATE OR REPLACE FUNCTION public.fn_minute_to_string(minute bigint)
   RETURNS text AS
 $BODY$DECLARE
 BEGIN
@@ -11,7 +11,8 @@ END
 $BODY$
   LANGUAGE plpgsql VOLATILE
   COST 100;
-ALTER FUNCTION public.fn_minute_to_string(integer)
+ALTER FUNCTION public.fn_minute_to_string(bigint)
   OWNER TO postgres;
-COMMENT ON FUNCTION public.fn_minute_to_string(integer) IS '分か60進文字に変換します
+COMMENT ON FUNCTION public.fn_minute_to_string(bigint) IS '分か60進文字に変換します
 ';
+
